requirements.txt
discord-py-interactions==5.15.0
aria2p==0.12.1
python-dotenv==1.1.1
jsonschema==4.25.1
jsonschema-gentypes==2.12.0 # for generating types from JSON schemas